requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.12
